+++ src/entities/balance.py
@@ -1,4 +1,5 @@
 from datetime import UTC, datetime
+from decimal import Decimal
 from uuid import UUID, uuid7
 
 from pydantic import Field
@@ -10,7 +11,7 @@ class Balance(BaseEntity):
 	"""Balance entity."""
 
 	id: UUID = Field(default_factory=uuid7)
-	amount: float = Field(description='Balance amount')
+	amount: Decimal = Field(description='Balance amount')
 	created_at: datetime = Field(
 		default_factory=lambda: datetime.now(UTC),
 		description='Balance created at',
